requirements.txt
c2pa-python==0.5.0
boto3
pyyaml
pyasn1==0.4.8
pyasn1-modules==0.2.7
flask
cryptography
arguably
python-pkcs11
